Move AppTheme to .theme package
