+++ app/src/main/java/dev/burnoo/demo/listapp/MainActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import dev.burnoo.demo.listapp.composable.AppRouter
-import dev.burnoo.demo.listapp.core.designsystem.AppTheme
+import dev.burnoo.demo.listapp.core.designsystem.theme.AppTheme
 import dev.burnoo.demo.listapp.di.WithDependencyInjection
 
 class MainActivity : ComponentActivity() {
