+++ core/designsystem/src/main/java/dev/burnoo/demo/listapp/core/designsystem/theme/Theme.kt
@@ -1,4 +1,4 @@
-package dev.burnoo.demo.listapp.core.designsystem
+package dev.burnoo.demo.listapp.core.designsystem.theme
 
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
